% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_queries.R
\name{get_burke_fellows}
\alias{get_burke_fellows}
\title{Make common soql queries from Salesforce}
\usage{
get_burke_fellows(cohorts = "current")
}
\arguments{
\item{cohorts}{Designations of the cohorts of fellows to query.
Accepts 'current', 'all', or a numeric vector.

Queries the City Forward Collective Salesforce Instance with
common soql queries.}
}
\description{
Make common soql queries from Salesforce
}
