% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_school_leaders.R
\name{get_school_leaders}
\alias{get_school_leaders}
\title{Retrieve list of School Leaders from City Forward Collective's Salesforce instance}
\usage{
get_school_leaders()
}
\description{
Retrieve list of School Leaders from City Forward Collective's Salesforce instance
}
