% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_burke_fellows.R
\name{get_burke_fellows}
\alias{get_burke_fellows}
\title{Retrieve list of Burke Fellows from City Forward Collective's Salesforce instance}
\usage{
get_burke_fellows(cohorts = "current")
}
\arguments{
\item{cohorts}{Designations of the cohorts of fellows to query.
Accepts 'current', 'all', or a numeric vector.}
}
\description{
Retrieve list of Burke Fellows from City Forward Collective's Salesforce instance
}
